model/script/rgbd/requirements.txt
# Base packages for RGBD YOLOv8
ultralytics>=8.3.112
torch>=1.7.0
torchvision>=0.8.1
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
PyYAML>=5.3.1
tqdm>=4.41.0
pillow>=7.1.2
seaborn>=0.11.0
scikit-learn>=0.19.2
setuptools>=65.5.1 